A mojolicious demonstration of Notice
This is an example of what mojolicious can do.
I managed to do in two days, with mojo, that I did in two weeks
with CGI::Application; that said, I managed to create a CGI::App
that performed the same task as a version that took me two years
two write from scratch.